Fix LSN not correctly updated
Before this change, the LSN indicated in the server heartbeat was not used to in the status reply, leading to the logical replication being delayed at least by 1 change. The issue did not have an impact on the data being received from the replication but it was an issue when trying to stop the PostgreSQL server, as it will wait for the replication to be up to date before shutting down.